pyproject.toml
[tool.poetry]
name = "grpc-xnet-socket-examples"
version = "0.1.0"
description = ""
authors = ["Alex Starche <alex.starche@ni.com>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
grpcio = "^1.45.0"
click = "^8.1.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
grpcio-tools = "^1.45.0"
mypy = "^0.942"
mypy-protobuf = "^3.2.0"
types-protobuf = "^3.19.15"
grpc-stubs = "^1.24.8"
protoletariat = "^0.7.5"
ni-python-styleguide = "^0.1.10"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 100
exclude = "xnetgrpc/"
